+++ assertions/src/main/java/internal/su/pernova/assertions/matchers/DecoratingMatcher.java
@@ -2,14 +2,13 @@ package internal.su.pernova.assertions.matchers;
 
 import static java.util.Objects.requireNonNull;
 
-import su.pernova.assertions.Description;
 import su.pernova.assertions.Matcher;
 
-public class IdentityMatcher extends DescriptiveMatcher {
+public abstract class DecoratingMatcher extends DescriptiveMatcher {
 
 	protected final Matcher delegate;
 
-	public IdentityMatcher(CharSequence description, Matcher delegate) {
+	public DecoratingMatcher(CharSequence description, Matcher delegate) {
 		super(description);
 		this.delegate = requireNonNull(delegate, "delegate is null");
 	}
@@ -17,10 +16,5 @@ public class IdentityMatcher extends DescriptiveMatcher {
 	@Override
 	public boolean match(Object actual) {
 		return delegate.match(actual);
-	}
-
-	@Override
-	public Description describe(Description description) {
-		return delegate.describe(description.appendText(" " + customDescription));
 	}
 }