+++ assertions/src/test/java/su/pernova/assertions/AssertionsTest.java
@@ -5,7 +5,12 @@ import static java.util.Collections.emptyList;
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 import static org.mockito.Mockito.mock;
 
+import static su.pernova.assertions.Assertions.assertThat;
+import static su.pernova.assertions.Matchers.does;
 import static su.pernova.assertions.Matchers.is;
+import static su.pernova.assertions.Matchers.match;
+import static su.pernova.assertions.Matchers.not;
+import static su.pernova.assertions.Matchers.regex;
 import static su.pernova.assertions.Subjects.subject;
 import static su.pernova.junit.jupiter.extension.Fixtures.forEach;
 
@@ -39,5 +44,15 @@ class AssertionsTest implements UtilityContractTest {
 	void noFailureWhenAssertionFailureThrowerDoesNotThrow() throws Exception {
 		fixture.setThrowers(emptyList());
 		assertDoesNotThrow(() -> su.pernova.assertions.Assertions.fail());
+	}
+
+	@Test
+	void negatedMatches() {
+		assertDoesNotThrow(() -> assertThat("apple", does(not(match(regex("[0-9]*"))))));
+	}
+
+	@Test
+	void emphasizedMatches() {
+		assertDoesNotThrow(() -> assertThat("123", does(match(regex("[0-9]*")))));
 	}
 }