+++ assertions/src/test/java/su/pernova/assertions/DefaultSubjectTest.java
@@ -5,8 +5,8 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import org.junit.jupiter.api.Test;
 
-import internal.su.pernova.assertions.matchers.IdentityMatcher;
-import internal.su.pernova.assertions.matchers.Is;
+import internal.su.pernova.assertions.matchers.IsMatcher;
+import internal.su.pernova.assertions.matchers.IsObject;
 import internal.su.pernova.assertions.subjects.DefaultSubject;
 
 class DefaultSubjectTest {
@@ -18,7 +18,7 @@ class DefaultSubjectTest {
 	@Test
 	void subjectMatchesNull() {
 		final DefaultSubject subject = new DefaultSubject(null);
-		assertTrue(subject.match(new IdentityMatcher("", new Is(null))));
+		assertTrue(subject.match(new IsMatcher("", new IsObject(null))));
 		subject.describe(description);
 		assertEquals(" subject", description.toString());
 		subject.describeMismatch(mismatchDescription);